requirements.txt
flask==1.0.3
flask-cors==3.0.8
flask-pymongo==2.3.0
flask-swagger==ui-3.20.9
flask_restful==0.3.7
tinys3==0.1.12
boto==2.49.0
